requirements.txt
Django
gunicorn
psycopg2-binary
celery
redis
setuptools
aiocontextvars
amqp
python-dotenv
dotenv
asgiref
billiard
CacheControl
cachetools
click
celery
certifi
cffi
chardet
click-plugins
click-repl
click-didyoumean
contextvars
Django
django-cors-headers
django-crontab
django-filter
django-logging-middleware
djangorestframework
djangorestframework-filters @ git+https://github.com/philipn/django-rest-framework-filters.git@master
djangorestframework-simplejwt
fcm-django
firebase-admin
grpcio
httplib2
idna
importlib-metadata
kombu
loguru
msgpack
packaging
proto-plus
protobuf
pyasn1
pyasn1-modules
pycparser
pyfcm
PyJWT
pyparsing
python-dateutil
pytz
redis
requests
rsa
six
sqlparse
typing-extensions
urllib3
gunicorn
wcwidth
zipp
